# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/project": {
            "post": {
                "description": "Create Project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Create Project",
                "parameters": [
                    {
                        "description": "Create project request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.SuccessWithDataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/res.ProjectResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.BadRequestError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.ForbiddenError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.ValidationFailedError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/project/user": {
            "get": {
                "description": "Project user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Project user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.SuccessWithDataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/res.ProjectUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.BadRequestError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.ForbiddenError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/project/users": {
            "get": {
                "description": "Project users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Project users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.SuccessWithDataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/res.ProjectUsersResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.BadRequestError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.ForbiddenError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "description": "UserProjects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.SuccessWithDataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/res.ProjectsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.BadRequestError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/errs.ForbiddenError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errs.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "errs.BadRequestError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "BAD_REQUEST"
                },
                "message": {
                    "type": "string",
                    "example": "Bad request."
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "errs.ForbiddenError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "FORBIDDEN"
                },
                "message": {
                    "type": "string",
                    "example": "Forbidden."
                },
                "status": {
                    "type": "integer",
                    "example": 403
                }
            }
        },
...
        "errs.ValidationFailedError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION"
                },
                "message": {
                    "type": "string",
                    "example": "Validation error."
                },
                "status": {
                    "type": "integer",
                    "example": 422
                }
            }
        },
        "project_dto.ProjectDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "test description"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "test"
                },
                "slug": {
                    "type": "string",
                    "example": "test"
                }
            }
        },
        "project_dto.ProjectUserDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "project": {
                    "$ref": "#/definitions/project_dto.ProjectDto"
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "user": {
                    "$ref": "#/definitions/user_dto.UserDto"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "req.CreateProjectRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 3
                },
                "slug": {
                    "type": "string",
                    "minLength": 3
                }
            }
        },
...
        "res.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errs.AppError"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "res.ProjectResponse": {
            "type": "object",
            "properties": {
                "project": {
                    "$ref": "#/definitions/project_dto.ProjectDto"
                }
            }
        },
        "res.ProjectUserResponse": {
            "type": "object",
            "properties": {
                "projectUser": {
                    "$ref": "#/definitions/project_dto.ProjectUserDto"
                }
            }
        },
        "res.ProjectUsersResponse": {
            "type": "object",
            "properties": {
                "projectUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/project_dto.ProjectUserDto"
                    }
                }
            }
        },
        "res.ProjectsResponse": {
            "type": "object",
            "properties": {
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/project_dto.ProjectDto"
                    }
                }
            }
        },
...
        "res.SuccessWithDataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "user_dto.UserDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@john.com"
                },
                "fullName": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}