# errors/error_types.go
// Code generated by go generate; DO NOT EDIT.
package errs


type AlreadyExistsError struct {
	Status  int    `json:"status" example:"409"`
	Code    string `json:"code" example:"ALREADY_EXISTS"`
	Message string `json:"message" example:"Already exists."`
}

type IncorrectPasswordError struct {
	Status  int    `json:"status" example:"401"`
	Code    string `json:"code" example:"INCORRECT_PASSWORD"`
	Message string `json:"message" example:"Incorrect password."`
}

type UserNotFoundError struct {
	Status  int    `json:"status" example:"400"`
	Code    string `json:"code" example:"USER_NOT_FOUND"`
	Message string `json:"message" example:"User not found."`
}

type UnauthorizedError struct {
	Status  int    `json:"status" example:"401"`
	Code    string `json:"code" example:"UNAUTHORIZED"`
	Message string `json:"message" example:"Unauthorized."`
}

type ForbiddenError struct {
	Status  int    `json:"status" example:"403"`
	Code    string `json:"code" example:"FORBIDDEN"`
	Message string `json:"message" example:"Forbidden."`
}

type TokenExpiredError struct {
	Status  int    `json:"status" example:"403"`
	Code    string `json:"code" example:"TOKEN_EXPIRED"`
	Message string `json:"message" example:"Token expired."`
}

type AuthError struct {
	Status  int    `json:"status" example:"400"`
	Code    string `json:"code" example:"AUTH_FAILED"`
	Message string `json:"message" example:"Failded authorize user."`
}

type ValidationFailedError struct {
	Status  int    `json:"status" example:"422"`
	Code    string `json:"code" example:"VALIDATION"`
	Message string `json:"message" example:"Validation error."`
}

type BadRequestError struct {
	Status  int    `json:"status" example:"400"`
	Code    string `json:"code" example:"BAD_REQUEST"`
	Message string `json:"message" example:"Bad request."`
}

type NotFoundError struct {
	Status  int    `json:"status" example:"404"`
	Code    string `json:"code" example:"NOT_FOUND"`
	Message string `json:"message" example:"Not found."`
}